Remove .tsx extensions. Resolve issues posted by Xander. FIxed google map container with containerStyle prop.
