+++ src/.eslintrc.json
@@ -1,14 +1,19 @@
 {
+  "extends": ["airbnb", "prettier", "prettier/react",
+    "plugin:prettier/recommended", "plugin:react/recommended"],
   "parser": "babel-eslint",
-  // "extends": "airbnb",
-  "env": {
-    "browser": true,
-    "node": true
+  "plugins": ["react", "prettier"],
+  "settings": {
+    "import/parsers": {
+      "@typescript-eslint/parser": [".ts", ".tsx"]
+    },
+    "import/resolver": {
+      "typescript": {}
+    }
   },
   "rules": {
-    "no-console": "off",
-    "comma-dangle": "off",
-    "react/jsx-filename-extension": "off",
-    "indent": [2, "tab"]
+    "react/jsx-filename-extension": [2, { "extensions": [".js", ".jsx", ".ts", ".tsx"] }],
+    "import/no-extraneous-dependencies": [2, { "devDependencies": ["**/test.tsx", "**/test.ts"] }],
+    "@typescript-eslint/indent": [2, 2]
   }
 }
