+++ src/.gitignore
@@ -7,3 +7,5 @@ yarn.lock
 .env
 assets/
 colors.js
+webpack2.config.js
+NOTES.txt
