+++ src/README.md
@@ -1,5 +1,3 @@
 # Harness
 
 ## Get Started
-
-yarn install
