+++ src/client/Components/DashboardTemp.jsx
@@ -1,9 +1,9 @@
-import React from 'react'
+import React from "react";
 
 export default function DashboardTemp() {
   return (
     <div>
       <h1>Coming Soon.</h1>
     </div>
-  )
+  );
 }
