+++ src/client/Components/SignUp.jsx
@@ -6,7 +6,6 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
@@ -133,10 +132,8 @@ export default function SignUp() {
         aria-labelledby="simple-modal-title"
         aria-describedby="simple-modal-description"
       >
-        
           <Typography variant='h4'>
             Registration Success! Please 
-            
           </Typography>
