+++ src/client/Components/Slider.jsx
