--- src/client/index.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './Components/App';
-
-ReactDOM.render(<App />, document.getElementById('root'));
