+++ src/package.json
@@ -8,15 +8,12 @@
 		"start": "npm run build && node server/index.js",
 		"client": "webpack-dev-server --mode development --devtool inline-source-map --hot",
 		"server": "nodemon server/server.js",
-		"dev": "concurrently \"npm run server\" \"npm run client\""
+		"dev": "concurrently \"npm run server\" \"npm run client\"",
+		"lint": "eslint . --ext .js,.jsx,.ts,.tsx"
 	},
 	"author": "Sean Haverstock",
 	"license": "ISC",
 	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^1.2.32",
-		"@fortawesome/free-regular-svg-icons": "^5.15.1",
-		"@fortawesome/free-solid-svg-icons": "^5.15.1",
-		"@fortawesome/react-fontawesome": "^0.1.13",
 		"@material-ui/core": "^4.11.0",
 		"@material-ui/icons": "^4.9.1",
 		"@material-ui/lab": "^4.0.0-alpha.57",
@@ -58,15 +55,19 @@
 		"concurrently": "^4.0.0",
 		"copy-webpack-plugin": "^5.0.3",
 		"css-loader": "^2.0.0",
-		"eslint": "^5.0.0",
-		"eslint-config-airbnb": "^17.0.0",
-		"eslint-plugin-import": "^2.11.0",
-		"eslint-plugin-jsx-a11y": "^6.0.3",
-		"eslint-plugin-react": "^7.7.0",
+		"eslint": "^7.16.0",
+		"eslint-config-airbnb": "^18.2.1",
+		"eslint-config-prettier": "^7.1.0",
+		"eslint-plugin-import": "^2.22.1",
+		"eslint-plugin-json": "^2.1.2",
+		"eslint-plugin-jsx-a11y": "^6.4.1",
+		"eslint-plugin-prettier": "^3.3.0",
+		"eslint-plugin-react": "^7.22.0",
 		"file-loader": "^3.0.0",
 		"html-webpack-plugin": "^3.2.0",
 		"mini-css-extract-plugin": "^0.6.0",
 		"nodemon": "^1.17.3",
+		"prettier": "^2.2.1",
 		"react-hot-loader": "^4.13.0",
 		"react-refresh": "^0.9.0",
 		"source-map-loader": "^0.2.4",
