+++ src/webpack.config.js
@@ -11,99 +11,97 @@ const outputDirectory = 'dist';
 const env = dotenv.config().parsed;
 
 const envKeys = Object.keys(env).reduce((prev, next) => {
-	prev[`process.env.${next}`] = JSON.stringify(env[next]);
-	return prev;
+  prev[`process.env.${next}`] = JSON.stringify(env[next]);
+  return prev;
 }, {});
 
-let isDevelopment = 'development';
+const isDevelopment = 'development';
 module.exports = {
-	mode: isDevelopment ? 'development' : 'production',
-	entry: './client/index.tsx',
-	output: {
-		path: path.join(__dirname, outputDirectory),
-		filename: './js/[name].bundle.js',
-	},
-	devtool: 'source-map',
-	module: {
-		rules: [
-			{
-				test: /\.(js|jsx)?$/,
-				exclude: /node_modules/,
-				use: [
-					// ... other loaders
-					{
-						loader: 'babel-loader',
-						options: {
-							// ... other options
-							plugins: [
-								// ... other plugins
-								isDevelopment && require.resolve('react-refresh/babel'),
-							].filter(Boolean),
-						},
-					},
-				],
-			},
-			{
-				test: /\.tsx?$/,
-				use: [
-					isDevelopment && {
-						loader: 'babel-loader',
-						options: { plugins: ['react-refresh/babel'] },
-					},
-					{
-						loader: 'awesome-typescript-loader',
-						options: { transpileOnly: true },
-					},
-				].filter(Boolean),
-			},
-			{
-				enforce: 'pre',
-				test: /\.js$/,
-				loader: 'source-map-loader',
-			},
-			{
-				test: /\.css$/,
-				use: [
-					{ loader: 'style-loader' },
-					{
-						loader: MiniCssExtractPlugin.loader,
-						options: {
-							publicPath: './Less',
-							hmr: process.env.NODE_ENV === 'development',
-						},
-					},
-					{ loader: 'css-loader' },
-				],
-			},
-			{
-				test: /\.(jpg|png|woff|woff2|eot|ttf|svg)$/,
-				loader: 'url-loader?limit=100000',
-			},
-		],
-	},
-	resolve: {
-		extensions: ['*', '.ts', '.tsx', '.js', '.jsx', '.json', '.jpg'],
-	},
-	devServer: {
-		port: 8050,
-		hot: true,
-		open: true,
-		historyApiFallback: true,
-		proxy: {
-			'/api': 'http://localhost:3000',
-		},
-	},
-	plugins: [
-		new CleanWebpackPlugin([outputDirectory]),
-		new HtmlWebpackPlugin({
-			template: './public/index.html',
-		}),
-		new webpack.DefinePlugin(envKeys),
-		new MiniCssExtractPlugin({
-			filename: './css/[name].css',
-			chunkFilename: './css/[id].css',
-		}),
-		// isDevelopment && new webpack.HotModuleReplacementPlugin(),
-		isDevelopment && new ReactRefreshWebpackPlugin(),
-	].filter(Boolean),
+  mode: isDevelopment ? 'development' : 'production',
+  entry: './client/index.jsx',
+  output: {
+    path: path.join(__dirname, outputDirectory),
+    filename: './js/[name].bundle.js',
+  },
+  devtool: 'source-map',
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)?$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            loader: 'babel-loader',
+            options: {
+              plugins: [
+                // ... other plugins
+                isDevelopment && require.resolve('react-refresh/babel'),
+              ].filter(Boolean),
+            },
+          },
+        ],
+      },
+      {
+        test: /\.tsx?$/,
+        use: [
+          isDevelopment && {
+            loader: 'babel-loader',
+            options: { plugins: ['react-refresh/babel'] },
+          },
+          {
+            loader: 'awesome-typescript-loader',
+            options: { transpileOnly: true },
+          },
+        ].filter(Boolean),
+      },
+      {
+        enforce: 'pre',
+        test: /\.js$/,
+        loader: 'source-map-loader',
+      },
+      {
+        test: /\.css$/,
+        use: [
+          { loader: 'style-loader' },
+          {
+            loader: MiniCssExtractPlugin.loader,
+            options: {
+              publicPath: './Less',
+              hmr: process.env.NODE_ENV === 'development',
+            },
+          },
+          { loader: 'css-loader' },
+        ],
+      },
+      {
+        test: /\.(jpg|png|woff|woff2|eot|ttf|svg)$/,
+        loader: 'url-loader?limit=100000',
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['*', '.ts', '.tsx', '.js', '.jsx', '.json', '.jpg'],
+  },
+  devServer: {
+    port: 8050,
+    hot: true,
+    open: true,
+    historyApiFallback: true,
+    proxy: {
+      '/api': 'http://localhost:3000',
+    },
+  },
+  plugins: [
+    new CleanWebpackPlugin([outputDirectory]),
+    new HtmlWebpackPlugin({
+      template: './public/index.html',
+    }),
+    new webpack.DefinePlugin(envKeys),
+    new MiniCssExtractPlugin({
+      filename: './css/[name].css',
+      chunkFilename: './css/[id].css',
+    }),
+    // isDevelopment && new webpack.HotModuleReplacementPlugin(),
+    isDevelopment && new ReactRefreshWebpackPlugin(),
+  ].filter(Boolean),
 };
